FEXCore: Moves BitUtils to FHU
No functional change
