+++ FEXCore/Source/Interface/Core/ArchHelpers/CodeEmitter/Emitter.h
@@ -4,12 +4,13 @@
 #include "Interface/Core/ArchHelpers/CodeEmitter/Buffer.h"
 #include "Interface/Core/ArchHelpers/CodeEmitter/Registers.h"
 
-#include <FEXCore/Utils/BitUtils.h>
 #include <FEXCore/Utils/CompilerDefs.h>
 #include <FEXCore/Utils/EnumUtils.h>
 #include <FEXCore/Utils/LogManager.h>
 #include <FEXCore/Utils/MathUtils.h>
 #include <FEXCore/fextl/vector.h>
+
+#include <FEXHeaderUtils/BitUtils.h>
 
 #include <aarch64/assembler-aarch64.h>
