+++ FEXCore/Source/Interface/IR/Passes/RegisterAllocationPass.cpp
@@ -13,7 +13,6 @@ $end_info$
 #include <FEXCore/IR/IREmitter.h>
 #include <FEXCore/IR/IntrusiveIRList.h>
 #include <FEXCore/IR/RegisterAllocationData.h>
-#include <FEXCore/Utils/BitUtils.h>
 #include <FEXCore/Utils/BucketList.h>
 #include <FEXCore/Utils/LogManager.h>
 #include <FEXCore/Utils/MathUtils.h>
@@ -24,6 +23,7 @@ $end_info$
 #include <FEXCore/fextl/unordered_set.h>
 #include <FEXCore/fextl/vector.h>
 
+#include <FEXHeaderUtils/BitUtils.h>
 #include <FEXHeaderUtils/TypeDefines.h>
 
 #include <algorithm>
